requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.21.0
bottle==0.12.21
bottle-websocket==0.2.9
cffi==1.15.1
cryptography==38.0.1
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
pefile==2022.5.30
pyAesCrypt==6.0.0
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
SQLAlchemy==1.4.41
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
